// packages/core/src/lib/games/DnD/types.ts
export enum EStats {
  Strength = 'strength',
  Dexterity = 'dexterity',
  Constitution = 'constitution',
  Intelligence = 'intelligence',
  Wisdom = 'wisdom',
  Charisma = 'charisma',
}

export enum ESkills {
  Acrobatics = 'acrobatics',
  AnimalHandling = 'animalHandling',
  Arcana = 'arcana',
  Athletics = 'athletics',
  Deception = 'deception',
  History = 'history',
  Insight = 'insight',
  Intimidation = 'intimidation',
  Investigation = 'investigation',
  Medicine = 'medicine',
  Nature = 'nature',
  Perception = 'perception',
  Performance = 'performance',
  Persuasion = 'persuasion',
  Religion = 'religion',
  SleightOfHand = 'sleightOfHand',
  Stealth = 'stealth',
  Survival = 'survival',
}

export enum EClass {
  Barbarian = 'barbarian',
  Bard = 'bard',
  Cleric = 'cleric',
  Druid = 'druid',
  Fighter = 'fighter',
  Monk = 'monk',
  Paladin = 'paladin',
  Ranger = 'ranger',
  Rogue = 'rogue',
  Sorcerer = 'sorcerer',
  Warlock = 'warlock',
  Wizard = 'wizard',
}

export enum ERace {
  Dragonborn = 'dragonborn',
  Dwarf = 'dwarf',
  Elf = 'elf',
  Gnome = 'gnome',
  HalfElf = 'halfElf',
  HalfOrc = 'halfOrc',
  Human = 'human',
  Tiefling = 'tiefling',
}

export enum EBackground {
  Acolyte = 'acolyte',
  Charlatan = 'charlatan',
  Criminal = 'criminal',
  Entertainer = 'entertainer',
  FolkHero = 'folkHero',
  GuildArtisan = 'guildArtisan',
  Hermit = 'hermit',
  Noble = 'noble',
  Outlander = 'outlander',
  Sage = 'sage',
  Sailor = 'sailor',
}